requirements.txt
aiogram==3.7.0
aiohttp==3.9.5
beautifulsoup4==4.12.3
lxml==5.2.1
python-dotenv==1.0.1
